backend/requirements.txt
# Backend-only requirements (for Docker/lean installs)
neo4j>=5.20.0
pyyaml>=6.0.1
structlog>=24.1.0
prometheus-client>=0.20.0

scikit-learn>=1.4.0
xgboost>=2.0.0
pyod>=1.1.0
pandas>=2.1.0
numpy>=1.26.0
joblib>=1.3.0

pyspark>=3.5.0
kafka-python-ng>=2.2.0

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.0

stable-baselines3>=2.3.0
sb3-contrib>=2.0.0
gymnasium>=0.29.0

tqdm>=4.66.0
